Replace in-memory store with PostgreSQL via pgx and sqlc
- Remove in-memory store; add PostgresStore with pgxpool
- Add sqlc-generated query layer and migration runner
- Add docker-compose.yml for local PostgreSQL
- Add CI workflow (go test ./...)
- Add EqualSplits, settlement tests, and index migration
- Update README with architecture and setup instructions

+++ migrations/004_create_expenses.sql
@@ -3,7 +3,7 @@ CREATE TABLE expenses (
     group_id UUID NOT NULL REFERENCES groups(id) ON DELETE CASCADE,
     paid_by UUID NOT NULL REFERENCES users(id),
     description TEXT NOT NULL,
-    amount_cents INTEGER NOT NULL CHECK (amount_cents > 0),
+    amount_cents BIGINT NOT NULL CHECK (amount_cents > 0),
     currency TEXT NOT NULL DEFAULT 'GBP',
     created_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
