+++ migrations/005_create_expense_splits.sql
@@ -1,6 +1,6 @@
 CREATE TABLE expense_splits (
     expense_id UUID NOT NULL REFERENCES expenses(id) ON DELETE CASCADE,
     user_id UUID NOT NULL REFERENCES users(id),
-    amount_cents INTEGER NOT NULL CHECK (amount_cents >= 0),
+    amount_cents BIGINT NOT NULL CHECK (amount_cents >= 0),
     PRIMARY KEY (expense_id, user_id)
 );
