+++ migrations/006_create_settlements.sql
@@ -3,7 +3,7 @@ CREATE TABLE settlements (
     group_id UUID NOT NULL REFERENCES groups(id) ON DELETE CASCADE,
     from_user_id UUID NOT NULL REFERENCES users(id),
     to_user_id UUID NOT NULL REFERENCES users(id),
-    amount_cents INTEGER NOT NULL CHECK (amount_cents > 0),
+    amount_cents BIGINT NOT NULL CHECK (amount_cents > 0),
     settled_at TIMESTAMPTZ,
     created_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
